The web engine is solved when changing game sides in analysis mode.
When picochess was in analysis mode and the web engine was also in, when pressing the dgt lever in web, the picochess analysis engine changed sides, but the web engine did not, it was constantly analyzing the side before the change. The same FEN is created on the web as in picochess to send it to the engine. Now both the picochess motor and the web coincide on the side when changing the upper clock lever.