The brave stage of removing the _engine_move_task that contained the call to engine.play(). Now there is no longer any play calls in the code.
At the same time there had to be added a coordination to prevent both sister classes ContinousAnalysis and PlayingContinousAnalysis to both ask for analysis at the same time. The old code was using pause_event to prevent Play and Analysis to be used at the same time for the same engine. Now that is performed using a new EngineLease class.
This is just for edge cases when we are switching between engine turn and user turn. Normally in most cases the PlayingContinousAnalysis is used when engine is thinking on its turn, and during user turn or pure analysis non-playing only ContinousAnalysis is used.